Create separate prod config; improve tree shaking
