+++ config/webpack.base.js
@@ -1,0 +1,33 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const path = require('path');
+const HtmlWebPackPlugin = require('html-webpack-plugin');
+
+const htmlPlugin = new HtmlWebPackPlugin({
+  template: './src/index.html',
+  filename: './index.html',
+});
+
+const config = {
+  entry: ['./src/index.tsx'],
+  output: {
+    path: path.resolve(__dirname, 'build'),
+    filename: 'bundle.js',
+  },
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js'],
+  },
+
+  module: {
+    rules: [
+      {
+        test: /.tsx?$/,
+        loader: 'ts-loader',
+        exclude: /node_modules/,
+      },
+    ],
+  },
+
+  plugins: [htmlPlugin]
+};
+
+module.exports = config;
