+++ package.json
@@ -5,9 +5,11 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "webpack serve --open --mode=development",
+    "dev": "webpack serve --open --env dev",
     "lint": "eslint --cache --ext .js,.ts,tsx . --fix",
-    "lint:no-fix": "eslint --ext .js,.ts,tsx ."
+    "lint:no-fix": "eslint --ext .js,.ts,tsx .",
+    "build": "webpack --env prod",
+    "analyze": "webpack --env prod --profile --json > build/stats.json && webpack-bundle-analyzer build/stats.json"
   },
   "repository": {
     "type": "git",
@@ -41,6 +43,7 @@
     "ts-loader": "^8.1.0",
     "typescript": "^4.2.4",
     "webpack": "^5.31.0",
+    "webpack-bundle-analyzer": "^4.4.0",
     "webpack-cli": "^4.6.0",
     "webpack-dev-server": "^3.11.2"
   },
