+++ src/components/Footer.tsx
@@ -1,4 +1,4 @@
-import { Container } from '@material-ui/core';
+import Container from '@material-ui/core/Container';
 import React from 'react';
 
 const Footer = () => {
