+++ src/pages/JobPage.tsx
@@ -1,15 +1,12 @@
-import {
-  /* Avatar, */
-  Box,
-  Card,
-  CardContent,
-  CardHeader,
-  Container,
-  Typography,
-} from '@material-ui/core';
+import Save from '@material-ui/icons/Save';
+import Box from '@material-ui/core/Box';
 import Button from '@material-ui/core/Button';
-import { Save } from '@material-ui/icons';
+import Card from '@material-ui/core/Card';
+import CardContent from '@material-ui/core/CardContent';
+import CardHeader from '@material-ui/core/CardHeader';
+import Container from '@material-ui/core/Container';
 import React, { ReactElement } from 'react';
+import Typography from '@material-ui/core/Typography';
 
 import { useParams } from 'react-router';
 import { useFetchJob } from '../api/useFetch';
